Feat/excel template (#50)
# Description
This is a new independent feature introduced to the DAMP lab
application.
The user is able to upload an excel file and do the following:
- Rename, Reorder, Delete and Add columns to an excel file.
- Save the changes as a template
- Apply the template to a new file

The above is done so that users can upload Abbott blood testing data and
transform the output and copy it to clipboard so that they can paste and
upload it into eLabs.

The technical changes is that a schema and resolvers were added to save
and manage templates.

## Checklist

- [yes ] This PR can be reviewed in under 30 minutes
- [yes] My code follows the style guidelines of this project
- [ yes] I have performed a self-review of my own code
- [ yes] I have commented my code, particularly in hard-to-understand
areas
- [ yes] I have made corresponding changes to the documentation
- [ yes] I have assigned reviewers to this PR.